# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "产线故障数据库后端 API",
    "version": "1.0.0",
    "description": "MongoDB 查询网关。默认地址为 http://127.0.0.1:18080；部署可通过 PORT 修改端口。当前服务不提供认证、授权或限流，生产环境应置于受信任网络或由上游网关保护。"
  },
  "servers": [
    { "url": "http://127.0.0.1:18080", "description": "本地默认地址" },
    { "url": "http://localhost:18080", "description": "本地别名" }
  ],
...
  "paths": {
...
    "/api/sync/incremental": {
      "post": { "tags": ["sync"], "operationId": "startIncrementalSync", "responses": { "202": { "description": "任务已启动", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncStatus" } } } }, "409": { "$ref": "#/components/responses/Conflict" }, "500": { "$ref": "#/components/responses/ServerError" } } }
    },
...
    "/api/sync/runs": {
      "get": { "tags": ["sync"], "operationId": "listSyncRuns", "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }], "responses": { "200": { "description": "同步运行历史" } } },
      "post": { "tags": ["sync"], "operationId": "startSyncRun", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["mode"], "properties": { "mode": { "type": "string", "enum": ["incremental", "full"] }, "taskIds": { "type": "array", "items": { "type": "string" } }, "startDate": { "type": "string", "format": "date", "description": "全量同步必填" }, "endDate": { "type": "string", "format": "date" } } } } } }, "responses": { "202": { "description": "编排运行已启动" }, "400": { "$ref": "#/components/responses/BadRequest" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
...
    "/api/sync/runs/{id}/retry": {
      "post": { "tags": ["sync"], "operationId": "retrySyncRun", "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "202": { "description": "重试运行已启动" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
...
  },
  "components": {
    "parameters": {
...
      "Id": { "name": "id", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 }, "description": "详情 ID。维修/视图优先使用列表返回的 id；订单通常为 source:AUFNR。" },
...
    },
    "responses": {
      "BadRequest": { "description": "请求参数无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "id is required" } } } },
      "NotFound": { "description": "记录不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "已有同步任务运行", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Error" }, { "type": "object", "properties": { "status": { "$ref": "#/components/schemas/SyncStatus" } } }] } } } },
      "ServerError": { "description": "数据库或服务端错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } }, "additionalProperties": true },
...
      "SyncStatus": { "type": "object", "required": ["state"], "properties": { "state": { "type": "string", "enum": ["idle", "running", "success", "failed"] }, "startedAt": { "type": "string", "format": "date-time" }, "finishedAt": { "type": "string", "format": "date-time" }, "message": { "type": "string" }, "summary": { "type": "object", "additionalProperties": true } } },
...
    }
  }
}